Refactor the database connection
