+++ .gitignore
@@ -5,3 +5,4 @@
 yarn.lock
 .DS_Store
 /dist/
+.vscode
